% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{apikey_from_sysenv}
\alias{apikey_from_sysenv}
\title{Get the API key from environment variable}
\usage{
apikey_from_sysenv(KEY)
}
\arguments{
\item{KEY}{A character string to get the relevant api.}
}
\value{
A character string with the API key, or error if not found
}
\description{
Get the API key from environment variable
}
\keyword{internal}
