% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{.collect_last_error}
\alias{.collect_last_error}
\title{Collect the most recent R error message using multiple strategies}
\usage{
.collect_last_error()
}
\description{
Tries, in order: rlang::last_error(), .Last.error (base R),
.Last.error condition message.
Returns NULL if nothing is found.
}
\keyword{internal}
