% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llmcoder-package.R
\docType{package}
\name{llmcoder-package}
\alias{llmcoder}
\alias{llmcoder-package}
\title{llmcoder: Coding Qualitative Data with Large Language Models}
\description{
This package facilitates the integration of large language models (LLMs), such as ChatGPT, into qualitative data processing and coding workflows. It provides wrapper functions to streamline interaction between data.frame objects—a standard data structure in R—and LLM-based APIs.
}
\author{
\strong{Maintainer}: Satoh Keiichi \email{keiichi.satoh@r.hit-u.ac.jp}

}
\keyword{internal}
