% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgets.R
\name{write_rprofile}
\alias{write_rprofile}
\title{Write options to ~/.Rprofile, replacing any previous llmcoder block.}
\usage{
write_rprofile(provider, model, api_key, ctx_lines, ollama_url, custom_url)
}
\description{
Write options to ~/.Rprofile, replacing any previous llmcoder block.
}
\keyword{internal}
